requirements.txt
pyobjc-core
pyobjc-framework-metal
pyobjc-framework-cocoa